+++ src/main.rs
@@ -1,5 +1,5 @@
 use std::{env, io};
 
 fn main() -> io::Result<()> {
-    editor_91to9::tui::run(env::args())
+    editor_91to9::cli::run(env::args())
 }